+++ src/scannerGet.java
@@ -1,0 +1,5 @@
+import java.util.Scanner;
+
+public class scannerGet {
+    static Scanner scanner = new Scanner(System.in);
+}
